fix compilation

+++ packages/Framework/Utils/Mesh3D/pycpp__init/mod.cc
@@ -159,7 +159,7 @@ public:
         for (std::int64_t iz = 0; iz < nz; ++iz) {
           double v = buf[ix*ny*nz+iy*nz+iz];
           if (v||include_empty)
-            printf("  cell[%li,%li,%li] = %g\n",ix,iy,iz,v);
+            printf("  cell[%li,%li,%li] = %g\n",(long)ix,(long)iy,(long)iz,v);
         }
 #endif
   }
